Update patch set 1
Patch Set 1:

I haven't reviewed this yet, but since I see Raphael just +2'ed it I thought I would add this comment right away :

IAndroidTarget is an interface. Since we're making changes to it (this, and the recent boot class path change etc), should we change it to be an abstract class?

That way we can add utility methods on it, as well as compatibility forwarders, and in general be more flexible with changing it in the future?

Patch-set: 1