Update patch set 1
Patch Set 1:

For a little context here, I'm on an ecryptfs home that slightly exacerbates the performance penalty of a large number of small write syscalls, so this makes a practical difference for my build times.

The performance difference of my test case on my ecryptfs partition is 6s => 3s, whereas it's only 3.5s => 3s for an unencrypted partition.  There's basically no drawback to patching this in, though.

Patch-set: 1