Update patch set 2
Patch Set 2:

Great, the tests look awesome!

I'm just waiting on confirmation from somebody involved with publishing validation to confirm what the specific checks are for application icons. I ran the check on various sample projects and found a bunch of violations, so I'm wondering if there are some exceptions to the rule (e.g. what if you specify a launcher activity and have an icon on that activity -- would that satisfy the publishing validation?).

Patch-set: 2