Update patch set 2
Patch Set 2:

Sure, i'd also like to have their feedback. Unless I missed something at that time, I've already tried to publish an application that was exactly in the situation you are describing. There was a launcher Activity with an icon attribute but no icon on the application tag. I was unable to upload it. Moving the icon from activity to application resolved the issue. 
What I am not too confident about, is that this kind of rule could change over time. So if you are able to have a list of publishing constraints that are unlikely to shift, i'd be happy to contribute.

Patch-set: 2