Update patch set 1
Patch Set 1:

(1 comment)

There's a test called tests/attrOrder already but it's a high level test. I want to add more localized test but I wanted to do it in another CL as I need to cherry-pick this one in r22, and it's pre-refactor so it's a pain.

Patch-set: 1
Label: SUBM=+1