Update patch set 2
Patch Set 2:

Xav would have a better sense of whether this is acceptable or not, but from my perspective, changing the signature from int -> long might break a bunch of users that just copy over the latest .jar.

A better approach might be to annotate the current method as deprecated and create a new method that takes in both a long timeout and a TimeUnit like the many methods in java.util.concurrent.

Patch-set: 2