Update patch set 1
Patch Set 1:

Yeah, except you have the gitignore info maintained in four different files in the template, some of which are redundant. Plus you're going to need to write code on the Java file anyway to not stomp on a file if it already exists. In the end it would be a lot less code and a lot easier to maintain to just implement it fully from Java.

Patch-set: 1