Update patch set 1
Patch Set 1:

To generate .hgignore we could just change the recipe copy line to make the filename .hgignore instead of .gitignore.

Again, I'm not disputing the fact that generating code from Java is simpler. But it's also known to be a bad way to do it, which is why servlets were replaced by jsp's (and ditto in other web frameworks).

We could address the duplication by allowing our templates to have a shared resource folder (we used to have one, where we hosted the v4-support lib jar before we had dependencies). That way the copy lines would copy from ${resources} instead of from the local template resources root. That's useful not just for this particular .gitignore issue, but for the many other duplicates resources (basically all the resources in NewAndroidApplication and NewAndroidLibrary are duplicated; I guess this is an unintentional side effect of splitting this template up into separate templates rather than keying everything off of a single ${isLibrary} variable and just copying conditionally in the recipe.xml.ftl file.

Patch-set: 1