Update patch set 6
Patch Set 6: Verified+1 Code-Review+1

In thinking about this, I realized that we should be flexible about how many numbers we require in a revision. So I've redone the revision system to be completely flexible.

With this overhaul, I'd appreciate it if you took a quick pass through

Patch-set: 6
Label: Code-Review=+1
Label: Verified=+1