Update patch set 1
Patch Set 1: Code-Review+2

This all LGTM, but I'm a little confused by the name "module". Isn't this for a new project? We don't create a settings.gradle file for modules right?

Perhaps we really need *another* template -- one for project (which puts *only* the project level things, e.g. the outer settings.gradle file and the outer empty build.gradle and the root .gitignore file), and then *in addition* a new module template (basically this one). The new wizard, when creating a new project would then instantiate both (as well as the optional activity one), but when adding a module, it would only do the module template inside the project dir.

This isn't necessary for *this* CL of course, but I think the terminology is a little misleading since NewApplication implied what it was (a project) and NewModule sounds like something smaller.

Patch-set: 1
Label: Code-Review=+2