Update patch set 2
Patch Set 2:

Is this something we only want for release? I would think things like checkTargetApi should be done all the time. Are the fatal checks configurable? (wondering if we can get to a point where there are too many of them?)

Also we probably want to disable it if we are running a full build with "build" or "check". We can query the taskgraph and disable the vitalLint task when the full lint task is detected.

Patch-set: 2
Label: Verified=0