Update patch set 2
Patch Set 2:

I can easily be persuaded that we should do it always; I tied it to release builds since that's the same thing we're already doing in ADT (there, it's done for Export APK only). I figured we wouldn't want to make debug builds slow (since those are iterative, and are done again and again). However, compared to dex'ing lint is probably really fast. Especially once the ECJ change goes in (right now, .java analysis is really slow from command line lint, but with ECJ that bottleneck is removed.)

Yes, the severity of issues is configurable, so if you use lint.xml to mark an issue fatal (or non-fatal) that should affect whether it's included in this check. Note to self: add a unit test for that.

Note that API check is *not* currently marked fatal. My rationale behind whether an issue is an "error" or is a "fatal error" is whether lint *knows* that the code in question is in fact an error. With API check, it's possible that the violating code is in code which is only reflectively called on the right platform etc, so I didn't "dare" to halt export if you simply haven't marked your code as @TargetApi(whatever).  Perhaps at a minimum we could add a sample showing how to set api check as fatal, to include it.

Also, I'd like to make it possible to configure severity directly from the gradle lintOptions, rather than requiring a lint.xml in the project. I'll take a look at that next (after fixing the jniLibs import issue.)

Patch-set: 2