Update patch set 4
Patch Set 4:

Hi, could you be please more specific regarding the verification date of this issue? It seems there is already ddmlib 22.5.0 in Maven central however this issue was not released with it. Why?

Please we are really depending on this one so it would be really awesome if things would be more smooth.

Patch-set: 4
Reviewer: Gerrit User 1035128 <1035128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0