Update patch set 1
Patch Set 1:

Yes, they can't edit the manifest merger unit tests from within Eclipse.

However, right now, opening the ADT project to edit *anything* results in the Problems view listing hundreds of errors. (That's how I ran into this; I went to update some lint related code.)

The right fix is to have a separate project for the unit tests. That's what we did for ADT itself (ADT's unit tests are in a separate project), but the tools base codebase is organized slightly differently.

It might be as simple as just adding the mockito dependency in there, but since I don't know much about how we build the production bits of ADT, I was worried of adding mockito into the IDE distro itself. 

Adding Siva to this in case he wants to chime in (he's working on updating the ADT build to use Tycho or something like that these days).

Patch-set: 1