Update patch set 2
Patch Set 2:

(1 comment)

This is a bit silly. I'd give it a -0.5 if I could.
Rather than do this, can we take the stance that Guava is accessible in all of our codebase? That would simplify usage. Eventually I got used to avoid Guava because I never know whether it's available or not in a given context. Can we just take one stance and stick to it?

Patch-set: 2
Reviewer: Gerrit User 1001100 <1001100@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0