Update patch set 1
Patch Set 1:

(4 comments)

while i disagree with the premise of the bug (67513), if we want to go this route and add a super simple activity, i'd just make sure to remove even more code so there's overlap between this and BlankActivity. we may also want to rename BlankActivity or this one.

Patch-set: 1