Update patch set 1
Patch Set 1:

(6 comments)

Generally this looks good. Jerome has some changes in the manifest task that is ready to go out almost, so you'll need to rebase when this goes in.

If you could also write a small blurb in base/build-system/changelog.txt that would be good. thanks.

Patch-set: 1
Label: Code-Review=0