Update patch set 2
Patch Set 2: Code-Review-1

Uh oh. We have to make sure that none of this code is run as part of library code in IntelliJ, where we don't have Guava 17 yet and some of those methods don't exist. We have to think about how to fix this. We can push for Guava 17 in IntelliJ 14, but we have to work in IntelliJ 14 too.

Patch-set: 2