Update patch set 1
Patch Set 1:

If we're going to do it, we should definitely do it ASAP (as in the next version of the plugin). But I think the android plugins might be core enough (and used by a high enough proportion of Gradle users) to get the same treatment as say the maven plugin or the ear plugin.

Patch-set: 1
Reviewer: Gerrit User 1004689 <1004689@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0