Update patch set 3
Patch Set 3:

Just a quick meta-comment: If you can, first submit a separate CL that has only formatting fixes for all the files. Then rebase this change on top of that.

If doing so is too cumbersome, then don't worry about it, at a quick glance, this didn't seem to have that many formatting changes anyway.

Patch-set: 3
Label: Code-Review=0