Update patch set 1
Patch Set 1:

(1 comment)

sorry for reviewing this so late, I was on vacation and missed it.

I think we should have a test project that set this up. we have the rsSupportMode that builds already, we should add it to the list of DeviceTest.

Patch-set: 1
Reviewer: Gerrit User 1001414 <1001414@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0