Update patch set 3
Patch Set 3: Verified+1

I tried this on a Windows 7 machine over the weekend, and we can delete the files after unmapping them, without shutting down the JVM.

I'm not convinced we need a test for running out of address space. On Windows 32-bit JVMs, the contiguous address space requirement means the JVM can't use more than ~1400Mb (-Xmx). If the mmap fails because it's too big, we wouldn't have enough heap to construct the in-memory representation of the snapshot anyway. If this is a problem, we can add code to detect this scenario in Android Studio and warn the user to switch to 64-bit.

Patch-set: 3
Label: Verified=+1