Update patch set 21
Patch Set 21: -Code-Review

(2 comments)

Ok so I verified that this works with the latest non released aapt. However we cannot check this in until we have updated the build server with a pre-release aapt.

In the mean time you're going to need to update the buildToolsRevision in the test project.

Patch-set: 21
Label: Code-Review=0