Update patch set 1
Patch Set 1:

This change will mean that the missing densities are exposed in places like resource chooser dialog. Developers might take that as a sign to add resources for this density. We may want to add a flag to suppress them in the UI.

Patch-set: 1
Reviewer: Gerrit User 1017685 <1017685@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0