Update patch set 2
Patch Set 2:

(1 comment)

I'm not sure this is the right way to do this. We should do this dynamically when the task is configured, and just enforce 7+ if API is 21+.

I don't think we need to break if the user explicitly says 6 but use 21. We can just output a warning or something.

Patch-set: 2
Label: Code-Review=0