Update patch set 2
Patch Set 2: Code-Review-2

Hi andrew, thanks for the contributions. This looks great. It's missing some tests but we can work through that.

However this is the wrong branch. You should contribute this to studio-1.0-dev instead.
(see our branches here: http://tools.android.com/build#TOC-Branches)
Note: this link shows that we would only accept contribution to our studio-1.1-dev branch but we were about to do this particular change anyway.

Which brings me to the other important point: This is an area where obviously we are currently working on. You were very close to do duplicate work on both your patches. It's critical that you talk to us first on the adt-dev google groups so that we can synchronize and avoid doing redundant work.

So first step: cherry-pick this into studio-1.0-dev, then we'll work on some details and the tests and we can get this in. Thanks again for the contribution.

Patch-set: 2
Reviewer: Gerrit User 1001414 <1001414@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2