Update patch set 1
Patch Set 1:

In the ADT version of this we had a common resource directory and the templates would look over in ../. I think it would be good to support this; I really hate all the code duplication between the various template files. But it will have to be post 1.0!

Patch-set: 1