Update patch set 2
Patch Set 2:

I know this is already submitted, but would like to ask about it.
Is there any reason not to return the model all the time?  If it's a performance issue (i.e. it evaluates the project again anyway), then  maybe it's better to have separate getModel method?  (I can do that in a different CL)  We can migrate the gradle-model tests into the integration test if we have that method.

Patch-set: 2
Reviewer: Gerrit User 1044651 <1044651@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0