Update patch set 6
Patch Set 6:

so, this is actually not great, because you don't see if the basic building of a project fails.

For instance in BasicTest, if the report fails, I think it would be good to know whether or not assembleDebug failed. Maybe we should keep an "empty" test in these projects.

Patch-set: 6
Reviewer: Gerrit User 1001414 <1001414@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0