Update patch set 1
Patch Set 1: Code-Review+1 Verified+1

Note: Most of the code in this CL was already reviewed and approved in this CL:
https://android-review.googlesource.com/112952

However, I reverted that CL later because it broke some sdklib tests, and (this late in the release) I was afraid to break the SDK manager.

In this CL I've included all the code again, but instead of changing the way FullRevision works, I've added a new subclass which implements the new semantics, and only lint uses that new subclass for now.

Patch-set: 1
Reviewer: Gerrit User 1004689 <1004689@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1
Label: Verified=+1