Update patch set 1
Patch Set 1: Code-Review+1 Verified+1

The Instantiator doesn't do everything (the class as is, fails the test without @GradleStyleSetters), it's part of the internal API and it's not under our control (we can benefit from changes to it until we bump up the required Gradle version).

This CL is a proof-of-concept, but the AST transform is 40 lines and allows us to drop a dependency on Gradle's internal API, and improve it (generate varargs settters etc.) easily.

Patch-set: 1
Label: Code-Review=+1
Label: Verified=+1