Update patch set 1
Patch Set 1:

Intellij should warn you (i.e. highlight in yellow) if you try to use the return value of a method that is marked as Nullable without checking for null.  I believe it only works in Java files though.  It might not be as relevant since this is an Impl class, but it's something to consider if we ever want to make a change like this.

Patch-set: 1