Update patch set 2
Patch Set 2: Code-Review-1

We need a 2nd flag in the resource merger. Even though we may want to disable the crunching of the png, we will always want to process the 9-patch(*)

(*)The current flag is mostly needed to disable the 9-patch processing that we do not want to do in library projects.

Patch-set: 2
Label: Code-Review=-1