Update patch set 1
Patch Set 1:

I had just merged this test yesterday. But I made a small edit from the gerrit UI to skip logging errors for @+id since it is expected that the resource will not be found (one has to be created for it). However, since I didn't run the tests after that minor change, it failed.

It would be a lesson for me to not use gerrit's edit feature.

Patch-set: 1