Update patch set 1
Patch Set 1:

(2 comments)

I'm wondering if we shouldn't change how we do things and instead have a single qualifier (that can handle language, language-region, or bcp-47), and encodes the value the same way as aapt does:

language,
optional region,
optional script.

This would allow us to do easier comparison, query, etc...

We'd need to change the parsing of the folder segments, and the UI maybe (since we have a single qualifier) but it would make things closer to what aapt does and make sure our selection of best matching resource work better.

I also feel this is kind of late for 1.1. Maybe we should talk about this tomorrow?

Patch-set: 1