Update patch set 1
Patch Set 1: Verified+1

(3 comments)

I tried this out and it worked for me. The Navigation Editor and Layout editors now 
both work with these templates. Layout Editor needed to prompt to add the missing
tools tag - which we could avoid be including it in the template - or not. Either 
is fine with me as all can be made to work from the UI now.

Patch-set: 1
Label: Verified=+1