Update patch set 1
Patch Set 1: Code-Review+2

Btw, we are reviewing the rest of the RPC and will be adding comments to the first CL where the whole code was checked in. Let us know if you prefer it being reviewed in a different way.

Patch-set: 1
Reviewer: Gerrit User 1047453 <1047453@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2