Update patch set 1
Patch Set 1:

We decided to create a brand new CL with all the code in RPC and aggregate all the separate reviews that we had done into this one. Note that this CL is based on one that deletes all the code. Seb, Ben, you can now take over, cherry-pick both CLs on your side and resolve the comments and upload the patches to this one. Once we are done, the two CLs can be squashed and submitted. Thanks!

Patch-set: 1
Label: Verified=0