Update patch set 4
Patch Set 4:

For AndroidManifest.xml setting, currently our sdk have no customized setup in our client jar. So here we're keeping the AndroidManifest.xml.ftl in this CL as it is now. We will be adding the setup into our client jar in our incoming release, and once it has the majority of users, we can change the setup through gradle.

For setRequestAgent in AdRequest, we're waiting for Orla official release so we can use it to track if the adRequest come from our template. I'll submit a new patch/CL for that.

Patch-set: 4
Label: Verified=0