Update patch set 1
Patch Set 1:

The overall goal is to try to be able to deprecate parts of sdklib that are there because of eclipse, or because of the standalone sdk manager.
The incremental goal is to reduce dependencies from the app on the inner workings of sdklib (the internal package), such that sdklib presents a sane interface, and the implementation of the internal packages can potentially be changed without affecting the users too much.
Since I'm not going to deal with eclipse/sdkuilib code, the current plan is to have two implementations of the internals, and the caller can specify if they want the new one, I guess.... probably the new implementation will also live in sdklib, and not depend on intellij (though that's not what I had at first hoped).
In this specific case, the point wasn't to remove the logger, the point was to remove the need for callers to use SettingsController, which is both internal and part of the eclipse/sdkuilib implementation. And since the callers I was interested in were using NullLogger anyway, this seemed like the logical approach. Furthermore, I was hope that the new implementation would be able to just generate loggers when needed and not have to pass them around all over the place... but I guess if the new implementation is going to live in sdklib, we can't use the intellij Logger and so that won't really work.
In the end, I'll put the logger back in for now, and have RemoteSdk create the settingscontroller if needed.

Patch-set: 1