Update patch set 1
Patch Set 1:

Yes, this is a new library, not a test for a library. (It's the library that lint unit tests use to easily check lint rules. Some 3rd party lint rule developers complained that they had to write their tests from scratch since the testing infrastructure was part of lint's own unit tests, not available as infrastructure they could depend on.

Actually, I don't think easymock is used by the infrastructure; I should be able to remove that. But junit is necessary.

Patch-set: 1