Update patch set 1
Patch Set 1:

Back then "Opening" a project did just that, opening a project, and if a .iml file was missing, the IDE will not see a module. Now that "Open" is really "Import" (for Gradle projects), .iml files will be generated all the time. 
IMHO, now it makes sense to ignore .iml files.

Patch-set: 1