Update patch set 7
Patch Set 7:

(7 comments)

Regarding AndroidLintInspectionToolProvider to make this lint rule show up as an inspection in Android Studio, yes, you have to check that in separately in the tools/adt/idea git project. The way to generate exactly the right contents to insert into the IDE is to run the AndroidLintInspectionToolProviderTest - it will fail when there's a new lint rule it doesn't know about and as part of its failure output it emits the right 3 pieces of code to insert.

Patch-set: 7
Label: Verified=0