Update patch set 4
Patch Set 4:

(4 comments)

Noted that I added system includes.  User probably will not be able to modify it directly, but it's needed to determine whether to "" or <> for include files.
I can't put that in NativeToolchain because the STL includes are system include, and they can be configured for each variant.

PTAL

Patch-set: 4
Label: Verified=0