Update patch set 5
Patch Set 5: Code-Review-1

Yes, but the part that says that was part of the current expansion didn't unmark the node when coming back from the "recursion". (Btw the linked algorithm just bails out if it finds a cycle). Mine and Chester solution fixed that, but you are correct and in the case you show it is now broken. So removing the +2 as we need to fix this properly.

Patch-set: 5
Label: Code-Review=-1