Update patch set 4
Patch Set 4:

(1 comment)

if this is causing trouble for you, we can go the java route and do something like "adb version" and "adb fullversion".

the problem we're trying to solve is that everyone's unhappy if we bump the protocol version unless there's an actual new feature, but we need to have some idea of what version someone's running when we get the bug report. the plan at the moment is -- where possible -- to not increment the version number, but to have the *devices* report (in their banner strings) the extra features they support. an old adb will just ignore this, but a newer adb can say "oh, you support larger blocks? great! i'll send you larger blocks..." or whatever.

Patch-set: 4
Label: Verified=0