Update patch set 4
Patch Set 4:

I think this means that the older version of tools will not work with this newer format, which might actually be a bit of a pain.

Xav had a good suggestion: if you could change the output in a way that doesn't spoil this parsing (like having it in a 2nd line, or somewhere before x.y.z in the first line), that would probably be better?

Patch-set: 4