Update patch set 6
Patch Set 6:

(1 comment)

It may help if you understand the model I have in mind for this API:

The resources are hierarchical and no level should care about other levels, except may be a black box holder for the resources/values inside it. A style resource value doesn't care about the items in it, and stores them as is, without looking in a list. The item resource value doesn't care about the value and just stores it, without looking in a String (hence no isFrameworkValue).

Patch-set: 6