Update patch set 1
Patch Set 1:

(1 comment)

> (1 comment)
 > 
 > Can you cover the DSL with unit tests, like we do in
 > AppPluginDslTest? I think we can get away without full integration
 > tests.
I don't think there is much benefit.  AppPluginDslTest is really an integration test disguised as a unit test.  It sets the DSL and then calls createAndroidTasks, which is really a workaround to the lack of public api to manual invoke evaluate on a Gradle project.
There isn't a simple way to invoke evaluate for the new plugin and the risk of error is high.

Patch-set: 1
Label: Verified=0