Update patch set 1
Patch Set 1:

> Using an empty collection is the Right Thing To Do - why didn't we
 > change Studio?

Mostly because there are some contrived scenarios where someone can be using an empty list of filters and it becomes a potentially breaking change...
Actually...Jerome, would it be better to break the backward compatibility?  The only place I know is when AS runs an apk.  So this change would allow user to deploy to more devices than they used to.  The risk is that there are other scenarios that I do not know about.

Patch-set: 1