Update patch set 1
Patch Set 1:

I was under the impression there are only dynamic proxies on the AS side and code executes in gradle JVM. Isn't it the case that some of the model implementation is actually groovy code?

Plus, this change is self-contained , it works as-is with no changes on studio side.

Patch-set: 1