Update patch set 1
Patch Set 1:

A warning at the top of the file won't show up in gerrit for review, if you only change one line in build.gradle. This is why I wanted to "make it harded" and make everyone stop for a moment to consider if the change is safe.

Patch-set: 1