Update patch set 1
Patch Set 1:

Unreachable == will be GC'ed, correct? If so, I'm not sure if there's any practical value for also including the unreachable instances.

As for AS, I'm probably erroneously including the unreachable instances in the display. I'll write myself up a bug for it. Thanks for point it out. :)

Patch-set: 1