Update patch set 10
Patch Set 10: Code-Review+1 Verified+1

For several tasks (e.g. ProcessAndroidResources) I want to be able to insert a rewrite step that points back to the original resources as they were merged, rather than the current comments in files solution.

This also means that we don't have to run all the parsers for all the different tools and can choose based on versions of tools for when the tools change to have more structured output.

This CL doesn't change the output at all, it just means we're not trying to match every line of output against multiple regular expressions. I am working on CLs based on this to do the merge persist and logging in a more generally and robust way.

Patch-set: 10
Label: Code-Review=+1
Label: Verified=+1