Update patch set 6
Patch Set 6:

Using the lower level api gives much more control and allow changes to be made without breaking backward compatibility

These serialises accept several variations of the json output, so they can parse the output from older gradle plugins when used by studio (e.g. source file positions can either be an array of objects or just an object or in the legacy way provided as two separate fields) 

It is also simple to define sensible defaults for when fields are omitted (e.g. in SourcePosition if the end values are omitted, they default to the start values)

Patch-set: 6