Update patch set 1
Patch Set 1:

I'm not sure I agree with this. The warnings when a getter is marked with @Input etc. and is not used should probably be explicitly suppressed for cases when gradle's view of the inputs is different than the task's.

Patch-set: 1
Reviewer: Gerrit User 1056183 <1056183@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0