Update patch set 4
Patch Set 4: Code-Review+1

(9 comments)

Talked to Esteban that we should use float for time as long loses precision and the value would not be accurate.

Also talked to David that it would be little messy to call add method in the for loop, so separate the samples conversion and the samples adding.

Patch-set: 4
Label: Code-Review=+1