Update patch set 1
Patch Set 1:

Thank you for contributing to Android Studio.
We hear the issue of finding and deleting the string and menu.

I was actually thing of a different solution to the same problem:
Instead of adding 2 checkboxes we could simply not generate the string and the menu when the activity is added to an existing project, but do generate them when the user starts with "New Project..."
(We like to limit the number of decisions on the new blank activity panel).

Do you see an issue with this alternate approach?
We intend to include such a fix some time in the near future.

Patch-set: 1
Reviewer: Gerrit User 1068658 <1068658@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0