Update patch set 1
Patch Set 1:

Glad you're aware that it's a pain point! :) 

It would partly resolve the problem I suppose, but we're then making the assumption that users are adding Activities to projects more often than starting a new project. For some cases I'm sure this is true, but for people who are working on a project with a single Activity and lots of Fragments or who often make a new project to try something out in a smaller, controlled setup, the issue would still be there.

I still believe that when initially creating the project it would be better to give users the choice I guess. Totally understand that you want to minimise the options on that page though. 

Another solution could be to have separate template? 'Default Activity' should have nothing generated, but a 'Default Activity with basics' (or something with a better name!) could have some other bits added? That would allow people to have a choice. What would be the default would be as you see fit.

I realize that you will want to minimise the number of templates because it's extra work to maintain, so perhaps a matter of weighing up the cost of the different options?

Patch-set: 1
Reviewer: Gerrit User 1078423 <1078423@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0