Update patch set 1
Patch Set 1:

Ahhh, excellent point, I had not thought of that - probably not as obvious for everyone, but I think that would indeed be a good enough compromise, at least for now.

My next question is obvious: when do you think you'll be able to do this? :D

As for this change request, it can probably be abandoned seeing as you have a better solution in mind. Thanks for the discussion!

Patch-set: 1