Update patch set 3
Patch Set 3: Code-Review+1

(3 comments)

Excellent, thank you so much for this contribution!

FYI, we're already well past feature freeze for 1.4, and we're doing stabilization, so this feature/check is too late for 1.4, and we don't have a branch created for 1.5 yet, so I'm going to sit on it for another week or two before submitting.

Patch-set: 3
Label: Code-Review=+1