Update patch set 2
Patch Set 2:

I don't know what happened to the master branch; it looks like it was reset as part of the studio-1.5 push and it's now pointing to an older snapshot - it doesn't even include the new stuff in 1.5. I assume that's why this is showing up as unmergeable.

Anyway, I've cherrypicked this into the internal master branch so it will show up in 2.0.0-alpha3. I'll also check with the build team what's going on with the branch such that we can push this into AOSP in the meantime as well until 2.0 is pushed.

(Oh, and as always, thanks for the contribution!!)

Patch-set: 2
Reviewer: Gerrit User 1004689 <1004689@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0