Update patch set 1
Patch Set 1:

(3 comments)

Hi Tenghui!  I can't seem to stay away from VectorDrawable : )

My use case is a small local tool that uses the code as a library, not the existing command line tool.  I noticed none of the existing parsing for conversion actually requires a File as input so it should take the more general purpose input stream.

This opens up new possibilities such as a pipeline where the svg is generated as output from another process in memory and never saved to disk.  Or input from over the network - it would make hosting the conversion in a web server easier (no need to write and manage request-lifetimed tmp files), though I'm not doing this yet.

Patch-set: 1
Label: Verified=0