Update patch set 1
Patch Set 1:

(5 comments)

Although, I'm not sure even in that case we would have a bug, because we're using DexFileNameSupplier to get the list of dex file names. But a failing test would convince me :)

Here are some comments to the code.

Patch-set: 1
Reviewer: Gerrit User 1104608 <1104608@85c56323-6fa9-3386-8a01-6480fb634889>