Added some general, non-Android related checks.
Adds a number of new checks that check for more basic coding issues.
These changes were prompted by the discussion at the following link
(https://groups.google.com/forum/#!topic/lint-dev/g63LZZclwTI).
They are:
- Checking for NullPointerExceptions.
- Checking for improper usage of the @NonNull annotation, including:
- Assigning a @NonNull variable to a nullable value.
- Returning a nullable value from a method annotated @NonNull.
- Passing a nullable value to a method when that parameter was
  annotated @NonNull.
- Checking for usage of null with Optionals, specifically, returning
a null from a method that returns an Optional, or assigning a nullable
value to a variable that is marked @NonNull.
- Checking for comparing two URLs using .equals().
- Checking for comparing variables of types that should be probably be
  compared using .equals() (such as a java.lang.Integer or an array)
  rather than ==.
- Checking for assignment to a variable in a condition (i.e., writing
  a = b instead of a == b).

The first three checks use a new class called NullInferer
(com.android.tools.lint.checks.NullInferer) which attempts to find
values that are nullable but are not marked as such (with @Nullable).

Change-Id: I66350714d015ac87eb2f37f75bc60a8b3522ad79