Update patch set 2
Patch Set 2: Code-Review+2

I've checked this into 3.2 in the internal repository. If you need this in AOSP in the meantime you can submit here as well; the CL will be replaced by the internal repo in master when we push the 3.2 sources to AOSP.

Thanks for the fix!

Patch-set: 2
Reviewer: Gerrit User 1004689 <1004689@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2